+++ packages/perseus-score/src/widgets/categorizer/validate-categorizer.test.ts
@@ -1,0 +1,31 @@
+import validateCategorizer from "./validate-categorizer";
+
+import type {PerseusCategorizerValidationData} from "@khanacademy/perseus-score";
+
+describe("validateCategorizer", () => {
+    it("tells the learner its not complete if not selected", () => {
+        const validationData: PerseusCategorizerValidationData = {
+            items: ["apples", "oranges"],
+        };
+
+        const userInput = {
+            values: [2],
+        } as const;
+        const score = validateCategorizer(userInput, validationData);
+
+        expect(score).toHaveInvalidInput("INVALID_SELECTION_ERROR");
+    });
+
+    it("returns null if the userInput is valid", () => {
+        const validationData: PerseusCategorizerValidationData = {
+            items: ["apples", "oranges"],
+        };
+
+        const userInput = {
+            values: [2, 4],
+        } as const;
+        const score = validateCategorizer(userInput, validationData);
+
+        expect(score).toBeNull();
+    });
+});