+++ packages/perseus/src/widget-ai-utils/radio/radio-ai-utils.ts
@@ -1,5 +1,5 @@
-import type {PerseusRadioUserInput} from "../../validation.types";
 import type radio from "../../widgets/radio/radio";
+import type {PerseusRadioUserInput} from "@khanacademy/perseus-score";
 import type React from "react";
 
 export type BasicOption = {