+++ packages/perseus/src/widgets/expression/expression.tsx
@@ -1,6 +1,15 @@
 import * as KAS from "@khanacademy/kas";
 import {KeyArray, KeypadInput, KeypadType} from "@khanacademy/math-input";
+import {
+    getDecimalSeparator,
+    type PerseusExpressionWidgetOptions,
+} from "@khanacademy/perseus-core";
 import {linterContextDefault} from "@khanacademy/perseus-linter";
+import {
+    scoreExpression,
+    type PerseusExpressionRubric,
+    type PerseusExpressionUserInput,
+} from "@khanacademy/perseus-score";
 import {View} from "@khanacademy/wonder-blocks-core";
 import Tooltip from "@khanacademy/wonder-blocks-tooltip";
 import {LabelSmall} from "@khanacademy/wonder-blocks-typography";
@@ -18,18 +27,10 @@ import {ApiOptions, ClassNames as ApiClassNames} from "../../perseus-api";
 import a11y from "../../util/a11y";
 import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/expression/expression-ai-utils";
 
-import getDecimalSeparator from "./get-decimal-separator";
-import scoreExpression from "./score-expression";
-
 import type {DependenciesContext} from "../../dependencies";
 import type {FocusPath, Widget, WidgetExports, WidgetProps} from "../../types";
-import type {
-    PerseusExpressionRubric,
-    PerseusExpressionUserInput,
-} from "../../validation.types";
 import type {ExpressionPromptJSON} from "../../widget-ai-utils/expression/expression-ai-utils";
 import type {Keys as Key, KeypadConfiguration} from "@khanacademy/math-input";
-import type {PerseusExpressionWidgetOptions} from "@khanacademy/perseus-core";
 import type {PropsFor} from "@khanacademy/wonder-blocks-core";
 
 type InputPath = ReadonlyArray<string>;