+++ packages/perseus/src/widgets/input-number/input-number.tsx
@@ -1,4 +1,10 @@
 import {linterContextDefault} from "@khanacademy/perseus-linter";
+import {
+    inputNumberAnswerTypes,
+    scoreInputNumber,
+    type PerseusInputNumberRubric,
+    type PerseusInputNumberUserInput,
+} from "@khanacademy/perseus-score";
 import {spacing} from "@khanacademy/wonder-blocks-tokens";
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
@@ -10,14 +16,8 @@ import SimpleKeypadInput from "../../components/simple-keypad-input";
 import {ApiOptions} from "../../perseus-api";
 import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/input-number/input-number-ai-utils";
 
-import scoreInputNumber, {answerTypes} from "./score-input-number";
-
 import type {PerseusStrings} from "../../strings";
 import type {Path, Widget, WidgetExports, WidgetProps} from "../../types";
-import type {
-    PerseusInputNumberRubric,
-    PerseusInputNumberUserInput,
-} from "../../validation.types";
 import type {InputNumberPromptJSON} from "../../widget-ai-utils/input-number/input-number-ai-utils";
 import type {PerseusInputNumberWidgetOptions} from "@khanacademy/perseus-core";
 
@@ -178,7 +178,7 @@ class InputNumber extends React.Component<Props> implements Widget {
     examples(): ReadonlyArray<string> {
         const {strings} = this.context;
         const type = this.props.answerType;
-        const forms = answerTypes[type].forms.split(/\s*,\s*/);
+        const forms = inputNumberAnswerTypes[type].forms.split(/\s*,\s*/);
 
         const examples = _.map(forms, (form) =>
             formExamples[form](this.props, strings),