+++ packages/perseus/src/widgets/numeric-input/numeric-input.tsx
@@ -1,5 +1,10 @@
 import {KhanMath} from "@khanacademy/kmath";
 import {linterContextDefault} from "@khanacademy/perseus-linter";
+import {
+    scoreNumericInput,
+    type PerseusNumericInputRubric,
+    type PerseusNumericInputUserInput,
+} from "@khanacademy/perseus-score";
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
 import _ from "underscore";
@@ -10,14 +15,8 @@ import SimpleKeypadInput from "../../components/simple-keypad-input";
 import {ApiOptions} from "../../perseus-api";
 import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/numeric-input/prompt-utils";
 
-import scoreNumericInput from "./score-numeric-input";
-
 import type {PerseusStrings} from "../../strings";
 import type {FocusPath, Widget, WidgetExports, WidgetProps} from "../../types";
-import type {
-    PerseusNumericInputRubric,
-    PerseusNumericInputUserInput,
-} from "../../validation.types";
 import type {NumericInputPromptJSON} from "../../widget-ai-utils/numeric-input/prompt-utils";
 import type {
     PerseusNumericInputWidgetOptions,