+++ packages/kas/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/kas
 
+## 0.4.11
+
+### Patch Changes
+
+-   Updated dependencies [[`9cabe689a`](https://github.com/Khan/perseus/commit/9cabe689a7aa143f95adf4556bf5c10d654a66ae)]:
+    -   @khanacademy/perseus-core@3.2.0
+
 ## 0.4.10
 
 ### Patch Changes